refactor: split lsp diagnostic types into microcrate
