+++ Cargo.toml
@@ -27,6 +27,7 @@ members = [
     "crates/perl-lsp-transport",
     "crates/perl-lsp-tooling",
     "crates/perl-lsp-formatting",
+    "crates/perl-lsp-diagnostic-types",
     "crates/perl-lsp-diagnostics",
     "crates/perl-lsp-semantic-tokens",
     "crates/perl-lsp-inlay-hints",
@@ -124,6 +125,7 @@ allow = [
     # Tier 3 — Analysis and indexing
     "perl-workspace-index",
     "perl-semantic-analyzer",
+    "perl-lsp-diagnostic-types",
     "perl-lsp-diagnostics",
     "perl-lsp-rename",
     "perl-lsp-code-actions",
@@ -257,6 +259,7 @@ perl-parser-core = { path = "crates/perl-parser-core", version = "0.10.0" }
 perl-lsp-transport = { path = "crates/perl-lsp-transport", version = "0.10.0" }
 perl-lsp-tooling = { path = "crates/perl-lsp-tooling", version = "0.10.0" }
 perl-lsp-formatting = { path = "crates/perl-lsp-formatting", version = "0.10.0" }
+perl-lsp-diagnostic-types = { path = "crates/perl-lsp-diagnostic-types", version = "0.10.0" }
 perl-tdd-support = { path = "crates/perl-tdd-support", version = "0.10.0" }
 perl-lsp-diagnostics = { path = "crates/perl-lsp-diagnostics", version = "0.10.0" }
 perl-lsp-semantic-tokens = { path = "crates/perl-lsp-semantic-tokens", version = "0.10.0" }
